+++ apps/storybook/stories/Button.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import "./button.css";
 
-export interface ButtonProps {
+export interface IButton {
   /** Is this the principal call to action on the page? */
   primary?: boolean;
   /** What background color to use */
@@ -16,13 +16,13 @@ export interface ButtonProps {
 }
 
 /** Primary UI component for user interaction */
-export const Button = ({
+export const Button: React.FC<IButton> = ({
   primary = false,
   size = "medium",
   backgroundColor,
   label,
   ...props
-}: ButtonProps) => {
+}) => {
   const mode = primary
     ? "storybook-button--primary"
     : "storybook-button--secondary";