+++ packages/utils/eslint.config.mjs
@@ -1,0 +1,2 @@
+import { config } from "@aristobyte-ui/eslint-config/react";
+export default config;