+++ packages/utils/icons/index.ts
@@ -1,0 +1,13 @@
+import { Copy } from "./Copy";
+import { Error } from "./Error";
+import { Info } from "./Info";
+import { Success } from "./Success";
+import { Warning } from "./Warning";
+
+export const Icons = {
+  Copy,
+  Error,
+  Info,
+  Success,
+  Warning,
+};