+++ packages/radio/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aristobyte-ui/radio
 
+## 1.0.21
+
+### Patch Changes
+
+- feat: add new package - utils
+
 ## 1.0.20
 
 ### Patch Changes